refactor : Protocol extension 구현
* checkPriority 메쏘드 추가
* CalculatorDelegate -> Calculatorable 리네이밍
* 인덴트 조정
* convertToComponentType 메쏘드 위치 변경
* 접근제어자 추가중
