--- Calculator/Calculator/Model/CalculatorDelegate.swift
@@ -1,17 +1,0 @@
-import Foundation
-
-enum CalculatorComponent {
-    case number
-    case `operator`
-    case equalSign
-    case allClear
-    case clearEntry
-    case signConverter
-    case dot
-}
-
-protocol CalculatorDelegate {
-    func convertToPostfixExpression(fromInfix input: [String]) -> [String]
-    func convertToComponentType(from userInput: String) throws -> CalculatorComponent
-    func calculatePostfixExpression(postfix: [String]) throws -> Double
-}
